# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "领域知识个性化生成与多智能体协同决策系统",
    "description": "领域知识个性化生成与多智能体协同决策系统后端API\n\n## 认证方式\n- 除 `/auth/login`、`/auth/register`、健康检查外，所有接口需在请求头携带 `Authorization: Bearer <access_token>`\n- Token 通过 `POST /api/v1/auth/login` 获取，有效期由 `JWT_EXPIRE_MINUTES` 控制\n- Token 过期后通过 `POST /api/v1/auth/refresh` 刷新\n\n## 统一响应格式\n所有接口返回 `{code, message, data, timestamp}` 结构，`code=200` 表示成功",
    "contact": {
      "name": "知识系统团队",
      "url": "https://github.com/knowledge-system"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "当前环境（相对路径）"
    },
    {
      "url": "http://localhost:8000/api/v1",
      "description": "开发环境"
    },
    {
      "url": "https://staging.example.com/api/v1",
      "description": "预发布环境"
    },
    {
      "url": "https://api.example.com/api/v1",
      "description": "生产环境"
    }
  ],
  "paths": {
...
    "/api/v1/learners": {
      "post": {
        "tags": [
          "学习者画像"
        ],
        "summary": "创建学习者画像",
        "description": "创建新的学习者画像\n\n- **user_id**: 关联用户ID\n- **real_name**: 真实姓名\n- **education_level**: 学历层次\n- **major**: 专业方向\n- **theoretical_foundation**: 理论基础(0-100)\n- 以及其他能力维度、知识盲区等",
        "operationId": "create_learner_api_v1_learners_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearnerProfileCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "学习者画像"
        ],
        "summary": "获取学习者列表",
        "description": "获取学习者列表，支持分页、搜索、筛选\n\n- **page**: 页码\n- **page_size**: 每页数量\n- **keyword**: 关键词搜索（姓名/专业/职位）\n- **education_level**: 学历过滤\n- **target_industry**: 目标行业\n- **learning_style**: 学习风格\n- **is_anonymized**: 是否脱敏",
        "operationId": "get_learner_list_api_v1_learners_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "education_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Education Level"
            }
          },
          {
            "name": "target_industry",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Industry"
            }
          },
          {
            "name": "learning_style",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Learning Style"
            }
          },
          {
            "name": "is_anonymized",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Anonymized"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/learners/{learner_id}": {
      "get": {
        "tags": [
          "学习者画像"
        ],
        "summary": "获取学习者详情",
        "description": "根据ID获取学习者详细信息\n\n- **learner_id**: 学习者ID",
        "operationId": "get_learner_detail_api_v1_learners__learner_id__get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "learner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Learner Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "学习者画像"
        ],
        "summary": "更新学习者画像",
        "description": "更新学习者画像信息",
        "operationId": "update_learner_api_v1_learners__learner_id__put",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "learner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Learner Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearnerProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "学习者画像"
        ],
        "summary": "删除学习者画像",
        "description": "删除学习者画像",
        "operationId": "delete_learner_api_v1_learners__learner_id__delete",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "learner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Learner Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/learners/batch-import": {
      "post": {
        "tags": [
          "学习者画像"
        ],
        "summary": "批量导入学习者",
        "description": "批量导入多组差异化学情数据\n\n- **learners**: 学习者列表",
        "operationId": "batch_import_learners_api_v1_learners_batch_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearnerBatchImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/learners/batch-export": {
      "post": {
        "tags": [
          "学习者画像"
        ],
        "summary": "批量导出学习者",
        "description": "批量导出学习者画像，支持JSON/CSV格式\n\n- **learner_ids**: 指定导出的ID，为空则导出全部\n- **export_format**: 导出格式(json/csv)\n- **include_sensitive**: 是否包含敏感数据",
        "operationId": "batch_export_learners_api_v1_learners_batch_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearnerBatchExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/learners/{learner_id}/analyze": {
      "post": {
        "tags": [
          "学习者画像"
        ],
        "summary": "学情分析",
        "description": "自动解析测试答题记录，提取理论强项、技能盲区\n\n- **learner_id**: 学习者ID",
        "operationId": "analyze_learning_api_v1_learners__learner_id__analyze_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "learner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Learner Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/learners/{learner_id}/anonymize": {
      "post": {
        "tags": [
          "学习者画像"
        ],
        "summary": "数据脱敏",
        "description": "一键对指定学习者敏感信息进行掩码处理\n\n- **learner_id**: 学习者ID\n- **fields**: 指定脱敏字段，为空则脱敏所有敏感字段",
        "operationId": "anonymize_learner_api_v1_learners__learner_id__anonymize_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "learner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Learner Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AnonymizeRequest"
                  }
                ],
                "title": "Anonymize Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/learners/{learner_id}/answers": {
      "post": {
        "tags": [
          "学习者画像"
        ],
        "summary": "添加答题记录",
        "description": "添加学习者答题记录（用于学情分析）",
        "operationId": "add_answer_record_api_v1_learners__learner_id__answers_post",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "learner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Learner Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRecordCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "学习者画像"
        ],
        "summary": "获取答题记录",
        "description": "获取学习者答题记录列表",
        "operationId": "get_answer_records_api_v1_learners__learner_id__answers_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "learner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Learner Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/learners/{learner_id}/blind-areas": {
      "get": {
        "tags": [
          "学习者画像"
        ],
        "summary": "获取知识盲区标签云",
        "description": "获取学习者知识盲区标签云数据",
        "operationId": "get_blind_areas_api_v1_learners__learner_id__blind_areas_get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "learner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Learner Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/report/learner/{learner_id}": {
      "get": {
        "tags": [
          "学情可视化报告"
        ],
        "summary": "生成完整学情报告",
        "description": "生成完整学情报告（前端可视化看板直接渲染）\n\n返回结构化数据：\n- 知识盲区热力图\n- 资源难度匹配曲线\n- 学习路径节点拓扑\n- 能力雷达图\n- 核心指标统计",
        "operationId": "generate_learner_report_api_v1_report_learner__learner_id__get",
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "learner_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Learner Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnonymizeRequest": {
        "properties": {
          "learner_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learner Id",
            "description": "学习者ID"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields",
            "description": "指定脱敏字段，为空则脱敏所有敏感字段"
          }
        },
        "type": "object",
        "title": "AnonymizeRequest",
        "description": "脱敏请求"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LearnerBatchExportRequest": {
        "properties": {
          "learner_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learner Ids",
            "description": "指定导出的学习者ID，为空则导出全部"
          },
          "export_format": {
            "type": "string",
            "title": "Export Format",
            "description": "导出格式(json/csv)",
            "default": "json"
          },
          "include_sensitive": {
            "type": "boolean",
            "title": "Include Sensitive",
            "description": "是否包含敏感数据",
            "default": false
          }
        },
        "type": "object",
        "title": "LearnerBatchExportRequest",
        "description": "批量导出请求"
      },
...
      "LearnerBatchImportRequest": {
        "properties": {
          "learners": {
            "items": {
              "$ref": "#/components/schemas/LearnerBatchImportItem"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Learners"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/LearnerBatchImportItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items",
            "description": "兼容旧版字段名"
          }
        },
        "type": "object",
        "title": "LearnerBatchImportRequest",
        "description": "批量导入请求"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}